Add support to neutral and reverse gears on DIRT 4 telemetry
This will set the correct values when the car is in neutral and reverse gears when using DIRT 4 on SimHub.

Thanks for your work on all the mods you created. I would most definitely not be playing GTA V if it wasn't for them. 
Also don't know if you changed something but the FFB on 4.7.2 is feeling great. 
Cheers