Change VehicleExtensions members to statics
Vehicle argument is always passed, so there's no real reason to have it as a class. Think the initial creation was just... a misguided/incomplete understanding.

This somewhat breaks other things that depend on it, but can be easily fixed by just mass-replacing the ugly g_ext. global with a VExt::